requirements.txt
Django
djangorestframework
django-cors-headers
mysqlclient
autopep8